requirements.txt
ansicon==1.89.0
arrow==0.16.0
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bitstring==3.1.7
blessed==1.17.10
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.7
django-picklefield==3.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jinxed==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
wcwidth==0.2.5
whitenoise==5.0.1
zope.interface==5.1.0
